First pass at generating sample data from openapi schema for json bodies
