+++ studio/src/pages/RequestorPage/store/utils.ts
@@ -1,8 +1,3 @@
-import { z } from "zod";
-import {
-  type KeyValueParameter,
-  enforceTerminalDraftParameter,
-} from "../KeyValueForm";
 import type { findMatchedRoute } from "../routes";
 import type { ProbedRoute } from "../types";
 import { type RequestMethod, type RequestType, isWsRequest } from "../types";
@@ -53,101 +48,6 @@ export function probedRouteToInputMethod(route: ProbedRoute): RequestMethod {
       return "HEAD";
     default:
       return "GET";
-  }
-}
-
-export function filterDisabledEmptyQueryParams(
-  currentQueryParams: KeyValueParameter[],
-) {
-  return enforceTerminalDraftParameter(
-    currentQueryParams.filter((param) => param.enabled || !!param.value),
-  );
-}
-
-export function extractQueryParamsFromOpenApiDefinition(
-  currentQueryParams: KeyValueParameter[],
-  route: ProbedRoute,
-) {
-  if (!route.openApiSpec) {
-    return enforceTerminalDraftParameter(currentQueryParams);
-  }
-
-  const parsedSpec = safeParseOpenApiSpec(route.openApiSpec, route.path);
-  if (!parsedSpec) {
-    return enforceTerminalDraftParameter(currentQueryParams);
-  }
-
-  // Extract query parameters from OpenAPI spec
-  const specQueryParams =
-    parsedSpec.parameters?.filter((param) => param.in === "query") ?? [];
-
-  // Convert OpenAPI params to KeyValueParameter format
-  const openApiQueryParams: KeyValueParameter[] = specQueryParams.map(
-    (param) => ({
-      id: param.name,
-      key: param.name,
-      value: param.schema.example?.toString() ?? "",
-      enabled: param.required,
-    }),
-  );
-
-  // Merge with existing parameters, preferring existing values
-  const mergedParams = openApiQueryParams.map((openApiParam) => {
-    const existingParam = currentQueryParams.find(
-      (p) => p.key === openApiParam.key,
-    );
-    return existingParam ?? openApiParam;
-  });
-
-  // Add any existing parameters that weren't in the OpenAPI spec
-  const additionalParams = currentQueryParams.filter(
-    (param) => !openApiQueryParams.some((p) => p.key === param.key),
-  );
-
-  return enforceTerminalDraftParameter([...mergedParams, ...additionalParams]);
-}
-
-const OpenApiParameterSchema = z.object({
-  parameters: z
-    .array(
-      z.object({
-        schema: z.object({
-          type: z.string(),
-          example: z.any().optional(),
-        }),
-        required: z.boolean(),
-        name: z.string(),
-        in: z.enum(["path", "query", "header", "cookie"]),
-      }),
-    )
-    .optional(),
-  requestBody: z
-    .object({
-      content: z.object({
-        "application/json": z.object({
-          schema: z.object({
-            $ref: z.string().startsWith("#/components/schemas/"),
-          }),
-        }),
-      }),
-    })
-    .optional(),
-  responses: z.record(z.string(), z.any()).optional(), // We don't validate responses structure since we only care about parameters
-});
-
-function safeParseOpenApiSpec(openApiSpec: string, routePath: string) {
-  try {
-    const spec = JSON.parse(openApiSpec);
-    const parsedSpec = OpenApiParameterSchema.safeParse(spec);
-    if (!parsedSpec.success) {
-      console.warn(
-        `Data in openApiSpec for ${routePath} was not in expected format. Here is the error: ${parsedSpec.error?.format?.()}`,
-      );
-      return null;
-    }
-    return parsedSpec.data;
-  } catch {
-    return null;
   }
 }
